:a: 新增加功能
1.通知的声音可以选择了
2.添加Email的功能
3.通知功能

+++ app/src/main/java/com/example/mengfei/todo/entity/ActionManager.java
@@ -2,16 +2,35 @@ package com.example.mengfei.todo.entity;
 
 import org.litepal.crud.DataSupport;
 
-/**
- * Created by mengfei on 2017/10/6.
- */
-
 public class ActionManager {
 
     public static void saveAction(Action action, String taskId) {
-        DataSupport.deleteAll(Action.class, "taskId=?", taskId);
         if (action != null) {
-            action.save();
+            int actionSumCount = DataSupport.where("taskId=?", taskId).count(Action.class);
+            if (actionSumCount == 0) {
+                action.save();
+            } else {
+                // 如果不为零，那么说明存在的那个Action可能是和这个要存储的Action是相同的
+                int actionCount = DataSupport.where("taskId=? and title=? and desc=? and type=?",
+                        action.taskId, action.title, action.desc, action.type).count(Action.class);
+                if (actionCount == 0) {
+                    //删除其他Action，保存我们想要保存的
+                    DataSupport.deleteAll(Action.class, "taskId=?", taskId);
+                    action.save();
+                }/*else {
+                 //如果不为零，那么说明数据库里面已经存在一个Action记录了，那么就没必要保存这个Action了
+                 }**/
+            }
+        } else {
+            //如果Action为空，那么说明这个Task没有设置Action，所以清除所有和Task有关的Action
+            DataSupport.deleteAll(Action.class, "taskId=?", taskId);
+        }
+    }
+
+
+    public static void deleteAction(Action action) {
+        if (action != null) {
+            DataSupport.deleteAll(Action.class, "taskId=?", action.taskId);
         }
     }
 }
